valid 체크 수정

+++ src/main/java/com/dot/tour_info_service_server/controller/FolderController.java
@@ -1,5 +1,6 @@
 package com.dot.tour_info_service_server.controller;
 
+import com.dot.tour_info_service_server.security.util.SecurityUtil;
 import com.dot.tour_info_service_server.service.CartService;
 import com.dot.tour_info_service_server.service.FolderService;
 import com.dot.tour_info_service_server.dto.CartDTO;
@@ -40,33 +41,42 @@ public class FolderController {
         return new ResponseEntity<>(response,HttpStatus.OK);
     }
 
-    //폴더 등록
+    //폴더 등록 - valid완료
     @PostMapping(value = "/register",produces = MediaType.APPLICATION_JSON_VALUE)
     public ResponseEntity<ResponseWrapDTO<Long>>register(@RequestBody FolderRegistDTO folderRegistDTO){
         ResponseWrapDTO response=new ResponseWrapDTO<>(false,null);
-        Long data=folderService.register(folderRegistDTO);
-        if(data>0){
-            response.setResult(true);
-            response.setData(data);
-            return new ResponseEntity<>(response,HttpStatus.OK);
+        if(SecurityUtil.validateMno(folderRegistDTO.getMno())) {
+            Long data = folderService.register(folderRegistDTO);
+            if (data > 0) {
+                response.setResult(true);
+                response.setData(data);
+                return new ResponseEntity<>(response, HttpStatus.OK);
+            }else{
+                return new ResponseEntity<>(response,HttpStatus.BAD_REQUEST);
+            }
         }
         return new ResponseEntity<>(response,HttpStatus.NOT_FOUND);
     }
 
-    //폴더명 수정
+    //폴더명 수정 - valid 완료
     @PutMapping(value = "/update",produces = MediaType.APPLICATION_JSON_VALUE)
     public ResponseEntity<ResponseWrapDTO<Long>>modify(@RequestBody FolderDTO folderDTO){
         ResponseWrapDTO response=new ResponseWrapDTO<>(false,null);
-        Long data=folderService.modify(folderDTO);
-        if(data!=-1l){
-            response.setResult(true);
-            response.setData(data);
-            return new ResponseEntity<>(response,HttpStatus.OK);
+        if(SecurityUtil.validateMno(folderDTO.getMno())) {
+            Long data = folderService.modify(folderDTO);
+            if (data != -1l) {
+                response.setResult(true);
+                response.setData(data);
+                return new ResponseEntity<>(response, HttpStatus.OK);
+            }
+            else{
+                return new ResponseEntity<>(response,HttpStatus.BAD_REQUEST);
+            }
         }
         return new ResponseEntity<>(response,HttpStatus.NOT_FOUND);
     }
 
-    //폴더 삭제
+    //폴더 삭제 -service에서 valid 완료
     @DeleteMapping(value = "/delete/{fno}",produces = MediaType.APPLICATION_JSON_VALUE)
     public ResponseEntity<ResponseWrapDTO<Long>> remove(@PathVariable Long fno){
         ResponseWrapDTO response = new ResponseWrapDTO(false,null);
@@ -79,30 +89,34 @@ public class FolderController {
         return new ResponseEntity<>(response,HttpStatus.NOT_FOUND);
     }
 
-    //스팟 등록
+    //스팟 등록 -value 완료
     @PostMapping(value = "/cart-append",produces = MediaType.APPLICATION_JSON_VALUE)
     public ResponseEntity<ResponseWrapDTO<Long>>spotAdd(@RequestBody CartDTO cartDTO){
+
         ResponseWrapDTO response=new ResponseWrapDTO(false,null);
-        Long data=cartService.addCart(cartDTO);
-        if (data!=-1l){
-            response.setResult(true);
-            response.setData(data);
-            return new ResponseEntity<>(response,HttpStatus.OK);
+        if(SecurityUtil.validateMno(cartDTO.getMno())) {
+            Long data = cartService.addCart(cartDTO);
+            if (data != -1l) {
+                response.setResult(true);
+                response.setData(data);
+                return new ResponseEntity<>(response, HttpStatus.OK);
+            }
         }
         return new ResponseEntity<>(response,HttpStatus.NOT_FOUND);
     }
 
-    //스팟 삭제
+    //스팟 삭제 - valid완료
     @DeleteMapping(value="/cart-delete",produces = MediaType.APPLICATION_JSON_VALUE)
     public ResponseEntity<ResponseWrapDTO<Long>>spotDelete(@RequestBody CartDTO cartDTO){
         ResponseWrapDTO response=new ResponseWrapDTO(false,null);
-        Long data=cartService.deleteCart(cartDTO);
-        if (data!=null){
-            response.setResult(true);
-            response.setData(data);
-            return new ResponseEntity<>(response,HttpStatus.OK);
+        if(SecurityUtil.validateMno(cartDTO.getMno())) {
+            Long data = cartService.deleteCart(cartDTO);
+            if (data != null) {
+                response.setResult(true);
+                response.setData(data);
+                return new ResponseEntity<>(response, HttpStatus.OK);
+            }
         }
-
         return new ResponseEntity<>(response,HttpStatus.NOT_FOUND);
     }
